+++ src/main/java/com/bigxodus/domains/Quest.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class Quest {
     private Long questId;
-    private Long latitude;
-    private Long longitude;
+    private Double latitude;
+    private Double longitude;
     private String title;
     private String description;
 }
